+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.4.0]
 ### Added
+- Add list trending (GitHub only)
 - Add list untagged stars
 - Add search string when listing stars
 - Add list stars by language and tag
@@ -34,7 +37,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-06-27
 ### Initial Release
 
-[Unreleased]: https://github.com/hoop33/limo/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/hoop33/limo/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/hoop33/limo/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/hoop33/limo/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/hoop33/limo/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/hoop33/limo/compare/v0.1.0...v0.2.0